moved hmm class to hmm.py.
